Adding support for BarResultRequest
