+++ Etabs_Adapter/Etabs_Adapter.csproj
@@ -199,6 +199,7 @@
     <Compile Include="Create\Bar.cs" />
     <Compile Include="Create\_Create.cs" />
     <Compile Include="Delete\_Delete.cs" />
+    <Compile Include="Read\Results\BarResults.cs" />
     <Compile Include="Types\NextIndex.cs" />
     <Compile Include="Read\Link.cs" />
     <Compile Include="Read\Material.cs" />
